requirements.txt
requests~=2.32.3
peewee~=3.17.8
pyTelegramBotAPI~=4.26.0
python-dotenv~=1.0.1
pydantic~=2.10.5
pydantic-settings~=2.7.1
chardet~=5.2.0
